+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-# 0.1.1
+# Changelog
+
+## 0.1.4
+- allow multiple subscribers to subscribe to a model
+- fix global subscriber
+- better error reporting
+
+## 0.1.1
 - make publish/broadcast a public method